Bug 700493: Search parent Resources, if not found in Local or Page
The spec mentions that the Page Resources are inherited by a Form
Xobject (not recommended), but Adobe and mupdf find, and use, the
Resource definition in a parent (ancestor) Xobject. Recursively
search the parent Resources ONLY if it isn't found in the Form's
Resources or the Page Resources (which may include Resources of
ancestors of the Page).