Avoid valgrind problems with copy_landscape.
The copy_landscape routine uses memflip8x8 to transpose 8 lines
of data to be 8 columns of data.

At the end of the line, unless the width is a multiple of 8, we
therefore end up accessing uninitialised data. This trips
valgrind.

We therefore introduce a special case variant for the end of the
line, where we mask off the undefined bits.

This can be seen by doing:

 make vg
 bin/gs -sOutputFile=test.ps -dMaxBitmap=400000000 -sDEVICE=ps2write
    -r300 -Z: -dNOPAUSE -dBATCH -K2000000 -dClusterJob
    /home/marcos/cluster/tests_private/comparefiles/Bug691740.pdf
 valgrind --track-origins=yes bin/gs -sOutputFile=/dev/null
    -dMaxBitmap=400000000 -sDEVICE=pkmraw -r300 -Z: -dNOPAUSE
    -dBATCH -K2000000 -dClusterJob test.ps