Bug 700512: Fix error from image_render_icc16.
If we render a skewed image using image_render_icc16, and the first
two pixels on the line have identical source values, the current
code would forget to do the color mapping. This would result in
a "Fatal" error coming back from the underlying pixel plot routine
which causes the whole program to exit.

The fix is simply to insist that the first pixel on a row must
always be mapped.