Bug 700306: Fix values returned by bbox device.
The "any part of a pixel" code in the new scan converter holds a
"cursor" position that determines where edges intersect with
the scanlines. The leftmost and rightmost positions of intersection
are held in this as values of type fixed.

The "null" cursor position is represented by using maximal/minimal
values. Unfortunately, I used {max,min}_int_in_fixed, rather than
{max,min}_fixed, which meant they weren't nearly as maximal/minimal
as they ought to have been.

This meant that the logic for inserting the real left/right hand
edges went wrong, and we output spurious positions into the
edgebuffer table. This in turn caused larger spans to be filled
to the underlying device than expected, hence upsetting the bbox.

For all other devices, there was no problem because overly large
or small values are clipped away.