Initial transparency support
This causes seg faults with some files on the cluster, so its disabled
by default, To enable set -dNOTRANSPARENCY=false.

Track the current media size in the pdf context, so that we can easily
use it (including selection of CropBox etc) without repeating the
search of the page dictionary. We need this as the default bbox for
transparency operations (especially page groups)

Twiddle with the currently unused transparency booleans.

Add pdf_trans.c/h to hold the code for dealing with transparency,
add these to the makefile

When processing q and Q, if the page uses transparency then we need
(apparently) to tell the PDF14 compositor about them.