txtwrite - fix use of incorrect index for recording Widths
When scanning multi-byte CIDFont text, we were using the text enumerator
'index' to store the widths of glyphs. However this is incorrect when
the CIDFont uses multiple bytes; the index is the byte count into the
input string, not the numebr of characters processed.

Use TextBufferIndex instead, which is correctly updated to track the
character index in the string, not the byte index.