XPS interpreter - address a valgrind problem
Bug #700557 "valgrind error pdfwriting text from an xps file"

The fault is not in pdfwrite, its the XPS interpreter. The pdfwrite
device copies a font, which involves calling gs_c_name_glyph() to
retrieve the name of a given glyph. The expectation (because this is
making a copy of a font) is that the returned string will be persistent
and not managed by the font or pdfwrite.

In the PostScript interpreter this is managed by adding the glyph name
to the PostScript name table, which is persistent to the end of job and
then freed.

Now, the XPS interpreter supplies a routine xps_true_callback_glyph_name
which is called to retrieve the glyph name. If there is a POST table in
the font, and it has names, and we can read it, we allocate some memory
and copy the name there before returning a pointer to that memory.

If there is a POST table, but its a format which has no names, or which
we cannot deal with, then the code simply made a buffer, printed a
dummy name into it, and returned that. The problem is that as soon as
we return, the buffer we created on the stack goes out of scope leading
to the valgrind error.

It appears that in XPS (and PCL where this code originated) there is no
mechanism like ht name table, and so even when we can find a name in the
POST table, we allocate memory for it, and allow it to leak!!

See the comment at around line 263 of xpsttf.c, and note that this is
the same code as in pl_glyph_name() in plfont.c

Since this memory leak already happens for the case where the POST
table can be read, I've (reluctantly) chosen to do the same in the case
where we simply invent a glyph (PCL doesn't do this). I'm not happy
about it, and the owners of the relevant interpreters really ought to
fix this properly IMO, or we should rewrite the 'copy_glyph' method
from the fonts so that it owns the memory for the name, but this would
be a huge undertaking.