Fix 700565 - SetHalftoneMethod parsing is wrong.
The parameter numbering was off by one and the check for object type
should not have exited the command.

Also change the operator to not report missing attributes.  This
results in changes to error test files in the Quality Logic suite,
however, we did not match the HP error report previously and don't
plan to do so.  Oddly, HP reports a C file and line number in the
report, we suspect this is a bug.