pdfwrite - use the correct enumerator when handling PCL bitmap fonts
Bug #700562 "Valgrind reveals use of unitialised value in text code"

This is a follow-on from commit c47512e5e638d903d69925f7ebab4de2aa3f481f
again what we have is confusion over *which* of the two enumerators we
should be using, when we are converting fonts which we can't preserve
into type 3 fonts.

For non-PCL cases we don't need the text enumerator for the 'default'
show functions, so we specifically set pdev->pte to NULL (at around line
3477 in gdevpdtt.c). However for PCL bitmap fonts, or the GL/2 stick
font, we do need the 'default' enumerator which we create in order to
process the 'BuildGlyph' for the character.

So we need to make sure we are using the correct enumerator when we try
to read back the 'returned_glyph'.