Workaround Shading/clist valgrind issues with trans_device.
When shading into the clist the code has to cope with 2 devices;
the target device, and the 'trans_device', which is updated to
use the source colorspace for the shading.

The number of components can vary between the two, meaning that
the shading code has to be passing (say) 1 component around,
but the call into the clist will attempt to write (say) 3 components.

The clist appears to cope with this mismatch on reading (and I
haven't found a case where we write too few components), so for
now, just add a hacky PACIFY_VALGRIND section to clear the
spare color components.