PDF interpreter - reset colour space after starting a Group
Bug #700572 "PDF to TIFF - only borders left after conversion - no errors"

The problem is that the PDF file does not set a fill (or indeed stroke)
colour space or colour before beginning to draw on the canvas. Thus it
relies upon the default colour values.

Because the page has a transparency Group, and the Group has a Colour
space, we set the colour space before we begin rendering. The defined
colour space of the group is an ICCBased space with four components.
The PDF Reference is clear that the default colour for an ICCBased space
is that all the components should be 0. This results in us drawing all
the filled elements in 0 0 0 0, or white.

We need to set the colour space before we define the group for the
transparency compositor, because the code relies upon the colour space
of the group being current at the time the group is 'defined'.

Grabbing the current colour space, before setting the group space, and
then setting it back again afterwards, effectively a gsave/grestore but
only for the colour space fixes this problem. This causes differences in
7 test files, which I believe are progressions.