Tweak cff_write_Encoding to avoid writing badly formed table.
The following:

 make gsdebugvg
 valgrind --track-origins=yes debugbin/gs -sOutputFile=/dev/null
     -dMaxBitmap=400000000 -sDEVICE=pdfwrite  -r72  -Z: -dNOPAUSE
     -dBATCH -K2000000 -dClusterJob Bug691693.pdf

results in valgrind warnings:

 Use of uninitialised value of size 8
    at 0x53F0B6: deflate_slow (in ghostpdl/bin/gs)
    by 0x53CDA5: deflate (in ghostpdl/bin/gs)
    by 0x6CB1B7: s_zlibE_process (in ghostpdl/bin/gs)
    by 0x6B35F0: swritebuf (in ghostpdl/bin/gs)
    by 0x6B31EB: s_process_write_buf (in ghostpdl/bin/gs)
    by 0x6B28E5: spputc (in ghostpdl/bin/gs)
    by 0x6B2CE1: sputs (in ghostpdl/bin/gs)
    by 0x7E3A43: type2_put_hintmask (in ghostpdl/bin/gs)
    by 0x7E4704: psf_convert_type1_to_type2 (in ghostpdl/bin/gs)
    by 0x7D9770: cff_put_CharString (in ghostpdl/bin/gs)
    by 0x7DAB9B: cff_write_CharStrings (in ghostpdl/bin/gs)
    by 0x7DC58B: psf_write_type2_font (in ghostpdl/bin/gs)
  Uninitialised value was created by a stack allocation
    at 0x7DAF8E: cff_write_Encoding (in ghostpdl/bin/gs)

Investigation shows that cff_write_Encoding is outputting a table
with uninitialised bytes in that. Not only that, in such cases
it can clearly omit some valid data.

Conceptually, the routine starts with an empty index table, with
uninitialised values. For each of the possible 256 entries within
the font, it attempts to encode the char to get the glyph index.
If this fails, then the glyph is not encoded, so not mentioned in
the table.

Assuming it is encoded, it finds the index for that glyph within
the subset. If there is no index, then it is not in the subset, so
no entry in the table.

If there is an index, we look up the glyph name. If there is no
glyph name, we choose not to encode the glyph in the table. This
is the (first) case which is causing us problems.

Finally, we check to see if the encoded index fits into the simple
(0..255) table without clashing with a previously encoded glyph;
if so, we insert it into the index table (incrementing the number
of entries we have inserted), if not, we have to resort to using a
supplement entry.

The code then assumes that the first 'actual_enc' bytes of index
make up the table. This is sadly not the case.

In the test given above, this routine is called several times, one
of which demonstrates that this assumption is false.

While the entries do yield a set of index values that form a prefix
within the table (from 0 to 20), number 14 has no glyph name. This
results in byte 14 of the table having a garbage value in, and
(more worryingly), because actual_enc is only incremented to 20,
we don't write the final entry of the table at all, effectively
dropping a glyph.

It is possible that something has gone wrong earlier in the process
and that this should be a fatal error, but observation of the code
shows that if we use a supplement, because a j maps to a previously
used i, we'll similarly miscount the length of the table.

A better solution, seems to me, to be to keep track of the maximum
encoded value that has gone into the index table, and to ensure
that every value within that table has a defined value.

The table is initialised with 0's, as Chris informs me that the CIF
SID for .notdef is 0.

We make matching changes to cff_Encoding_size.