Undef check_r6_password and PDFexecform
Immediate binding and undef for check_r6_password.

Move PDFexecform from gs_lev2.ps to pdf_draw.ps (the only place it was
used), and put it and .PDFexecform in pdfdict.  PDFexecform and
.PDFexecform don't need odef -- just make them executeonly (even though
they will be hidden in pdfdict).

Also change the initial size of pdfdict to just over the actual size
(it was growing to > 400 maxlength).