Bug 700533: Undo FT over/underflow workaround
Older versions of Freetype had a problem with overflow (and underflow) in the
Truetype hinting code. That was resolved some time ago with the addition of
(or handling of) the F_dot_P value.

In the file associaited with this issue, the glyph offset is caused by this
workaround.

Thus the workaround is no longer required, and as it actually causes a problem,
it is hereby removed.