Undef .set/.currentfilladjust, re-def .set/.currentfilladjust2
The .setfilladjust, .currentfilladjust were just wrappers around the
'2' versions.  They were used in only a couple places, and were just
dupping/popping the extra args, so I replaced these with equivalent
uses of the underlying operators.

These were "compatibility" operators for .setfilladjust2 and
.currentfilladjust2, which do the same basic thing but take two values
(x,y) instead of 1.

Robin wants these to be official extensions because they are useful
for testing, so I have taken out the code that undef's .setfilladjust2
and .currentfilladjust2.  We need to add documentation for these.