Bug 700584:  cups device
The cups device has a pile of different color spaces.  It is up to the cups device to
ensure that the color is handled properly for these.  RGBA should certainly be treated
as a 32 3 component color similar to our pngalpha device.  RGBW will be treated the same way.

Note that I looked at the output using Mike Sweet's rasterview program.  I suspect there
probably is some work that needs to be done on that viewer to ensure proper viewing of
this type of data.  With this patch gs is doing all it can.