Extend the transparency detection to optionally include spot colours
Seems the code expects us to supply the number of spot plates in a PDF
page as 'PageSpotColours' to the device.

Extend the transparency detection code so that if the device has a
PageSpotColours entry, we also pick up any spot colours (Separation or
DeviceN colour spaces) used on the page. If we are checking for spot
colours we also must not terminate the checking as soon as we find any
transparency, but we can stop any further transparency checks and simply
keep walking the Resources for more spot colours.

If we find any spots, use gs_putdeviceparams to send the number of
spots to the device; note this may cause the device to close, and need
to be reopened.