Undef .promptmsg in gs_init.ps
There used to be some code related to readline in lib/gs_rdlin.ps
that would redfine the .promptmsg, but that code is now gone.
There doesn't seem to be any reason to keep .promptmsg any more.
If somebody wants to redefine /prompt, that is still fine and is
documented in the PLRM.

Change to use immediate evaluation and then undef it.