Change the creation of all reference counted objects. Previously we had a central 'free' procedure for all reference counted objects, but they were created 'ad hoc' when required. This commit changes to a central 'alloc_object' function.
Similarly pdf_stream objects (which are not reference counted) are also
now created in a single function.

Cope with the page /Rotate key being greater than 270 degrees.

Due to recent changes in the interpreter API, delete the 'remove device'
function. Do a grestore (to match the gsave in pdf_impl_set_device
which is now called from pdf_impl_init_job) in pdf_impl_dnit_job so that
we don't corrupt the gstate stack.