Deep color transparency
Previously, all transparency blending has been done using the
pdf14 compisitor working in 8 bits. Here, we extend it to be
capable of working in 16 bits.

When the compositor is created, it looks at the underlying device;
if that device uses more than 8 bits per component (plus 8 bits
for tags, if tags are in use), then we use 'deep' (i.e. 16 bit
per component) pdf14 buffers.