Undef .vmreclaim in gs_init.ps
This was tricky because it can be used in showpage, which wasn't bound.
Previous commit makes it possible to bind showpage.

It's used for debugging in the memento build, but this is useful so I
needed to maintain that functionality.

Also used in gs_lev2.ps, so make sure that is bound.

Since this operator is defined in C in the level2dict, it needs to
be undef'd from both systemdict and level2dict.