Update psd devices to only do one pass while rendering.
Previously, PSD devices would loop over each component in the file
and then over each line in the file. For non-page mode rendering
therefore, we would render each page as many times as there were
components.

Change the logic here to loop over each component, then over each
line.

This means we seek within the file as we write, but we already
impose restrictions on the output for this device to ensure that
we don't write multiple pages to the same file, so this is no
great loss.