Bug 700606: Fix problem with interpolated indexed >8bit images.
In the initial decode of the line, all but the first pixel of the
line is getting a value 0.

We look to be wrongly applying the nth map function to the nth
pixel, when the nth map function is intended for the nth component
of the pixel. Given we are only processing indexed images at this
point, we therefore always want the 0th map function.