Undef .getpath in gs_init.ps
This was another case where the operator is defined in C and was put
into the level2dict for no apparent reason.  The code that uses it (in
pdf_ops.ps and pdf_draw.ps) didn't have level2dict available on the
dictstack when it defined the operators, so it wasn't getting bound.

Solution is to take .getpath out of level2dict.  Added zupath_op_defs
for it and fixed makefile.  Then can easily undef it.