Many changes for transparency detection and spot colour counting:
when REFCNT_DEBUG is true, store the context in each object, so we
can check if an object being freed is still in the cache.

Add pdfi_dict_knownget, pdfi_dict_knownget_type and pdfi_knownget_number
to simplify checking for existence and retrieval of objects, we often
want to get an object, and take action with the object, but only if it
exists.

Add code to detect an attempt to dereference a free object, and return
a null object (and emit a warning in debug mode).

Reformat lots of code to use these.

Relax all the checking for stream length, when dereferencing indirect
referencews. Iif we fail to find a well formed /Length key, if the
offset + Length exceeds the file size, if we fail to read an 'endstream
endobject' sequence, do not throw an error for any of these, just return
and hope that it all works out.....

Swap the order of freeing key/value pairs when freeing dictionary
contents. If there's a problem with the value, at least we can examine
the key to see what failed, if we do this.