Coverity ID 333614 sign extension
An intermediate calculation results in a cast to 'int' which is then
promoted (on 64-bit systems) to an unsigned long long. If the top bit
of the intermediate value is set, then all the top bits of the final
value will be set.

Do an explicit cast to an unsigned to avoid this.