Coverity ID 333611
The error handling code in the run length encode filter confuses
Coverity because it relies on a series of 'got' operations that jump
into a block of otherwise unreachable code.

The code is (IMO) ugly and confusing, so refactor it to avoid the
'if (0)' blocks. This results in many fewer goto labels, and simpler
code.