Coverity ID 33608
We weren't acting on a return value. If we get an error, which I believe
should actually be impossible in this case, we might potentially have
an invalid object being pointed to by an array.

Just in case, if we get an error here, discard the array we just
allocated and propagate the error.