Add final section that reports any whitelisted operators that are not seen.
By setting the whitelist entry to //false initially, and the //true when
the operator is encountered in the scan, we can report elements that are in
the whitelist that probably don't need to be there (they've already been
undef'ed).

Also for operators, since they may be encountered in a dict in "bare" form,
or may be seen in packedarray procedures with "--" in front and back, we
check for either having been encountered by pulling out the "bare" form
and checking if that whitelist entry has been seen.