Add a gsave/grestore pair to all group being/end operations.
This is required so that we can (now) set the current colour space to
the one in the Group. I'm not sure why this is required, since we pass
the colour space as one of the parameters, but the transparency code
expects to be able to pull information form the current colour space.

Explicitly set the opacity to 1 when starting a group.