Fix Bug #700575 valgrind problem with decode_char.
Previously PCL stored the character code when encoding and returned it
upon decode.  This was only implemented properly for TrueType fonts,
ie. Intellifont or bitmap fonts did not set the character which
resulted in an unitialized read.  Ostensibly (for PCL/XL) high level
devices pass the character back upon decode, so it is only necessary
to return the passed in value and we can dispense with storing the
last character.