Add a gsave/grestore pair around push/pop of transparency compositor. This resolves most of the transparency problems.
Fix a typo causing an assignment instead of a conditional.

reformat the code for pushing the transparency compositors so that if
we get an error we can abandon transparency for this page, but continue
to render the page without transparency.

This brings us to 1 seg fault with the simple page Group code.