When checking for an SMask in an image dictionary, if we don't get a dictionary back, then treat it as not being present. A file tried to use a free entry, which resulted in us getting a null object.
Remove the now erroneous warning that image SMasks are not supported
because it was confusing me :-)

Processing SMask entries means processing an image or form while before
we start the 'current' image or form. To make this easier break the
existing 'pdfi_Do' routine in half. The part that processes the stack
and retrieves the image dictionary from the name remains unchanged, the
remaining functionality becomes a new static routine
'pdfi_do_image_or_form'. We can then call this routine from
pdfi_do_image as well as form pdfi_Do.