Fill_stroke_path fixes for overprint and transparency and with clist
Add "color_is_stroke" to the gs_gstate so that overprint (in the pdf14
and overprint compositors) can select the "drawn_comps" appropriate for
the fill or stroke. The two values are saved in the compositor device.

The color_is_stroke is passed to the compositors as a parameter and is
written as one of the flags to the clist when the params are serialized.

The overprint device cannot use the gx_default_fill_stroke_path -- it
now has its own procedure that can manage the color_is_stroke value for
the fill and stroke parts of the operation.

In do_fill_stroke, we need to set up the stroke color first, THEN the fill
color so that we will have the fill color set for when we do the actual
call of fill_stroke_path.

We aren't sure if it is correct, but in pdf14_pop_transparency_group, we
use the bitwise OR of the stroke and fill drawn_comps. This may result in
some components being overwriten in the stroke (where it extends outside
the fill) if overprint_mode is 1 (OPM 1) and the fill had components that
were zero in the stroke color.

One obsolete comment in the overprint fill_in_procs about suppression of
overprint was removed since it has not been the case for a long time,

This commit gets rid of the clist cmd_opv_swapcolors since it is no longer
used.