Fix (non-PostScript) processing of -d switch with radix
The -d switch can take a radix as part of the value assigned to a
named switch. For example -dDisplayFormat=16#20808 is described in the
display device documentation as selecting CMYK rendering.

The implementation in plmain.c was completely broken and returned the
error "Malformed base value for radix" for any switch using '#'.

The first problem is that we increment the pointer 'v' after checking
'c' to see if its a number. That means when its not a number (ie #) we
have moved v on past the # character. So the test for *v being '#'
fails.

Later we use 'val' which has been initialised to the '#' symbol in the
string. We check to see its a number or A->F or a->f and if its not
then we throw an error. Of course its initially pointing at '#' so
it fails immediately.