Check error codes for configuration errors with the pdfwrite family
Bug #800819 "Incorrect parameter value for sColorConversionStrategy is silently ignored."

The low level parameter checking for the pdfwrite family wasn't
honouring error returns from certain kinds of parameters, most notably
those involving lists of permitted values (eg ColorConversionStrategy)

This commit throws an error for those cases.

This causes one file in the regression test suite 'pt.ps' to throw an
error with the ps2write device. This is because the PostScript program
contains an invocation of setdistillerparams which sets the
ColorImageFilter (and GrayImageFilter) to Flate instead of DCT. Level 2
PostScript doesn't include the Flate filter, that is a level 3 addition
so we now throw an error on that file.