pdfwrite - don't write an incomplete Outline entry
Bug #700816 "Valgrind shows problems with pdfmarks"

The file in question has a recursive Outline entry, which eventually
leads to an execstackoverflow. This causes the pdfwrite device to be
closed, which leads to it closing and writing the Outkine tree as part
of the cleanup. Because of the way the /OUT pdfmark works, this means
there is a dangling reference (because the last /OUT pdfmark had a
child, but we haven't yet received the child data).


There was in fact already code to deal with this, and not write out
the final outline entry. However it relied upon an uninitialised value
being 0, which of course it usually isn't.

Fixed by initialising the value to 0, which should avoid the valgrind
warning, and prevents us writing an invalid Outline tree entry as well.