Bug 700678: Check for embedded font when substituting
When we apply a font name mapping, make sure that the ultimate font we load
comes from disk, and is not embedded in the job.

To aid differentiating between a substitute CIDfont (cidfmap), a "real" CIDFont
resource, and an embedded CIDFont, add a /ResourcePath key to CIDFont resource
dictionaries (susbsitute CIDFonts contain /Path key).