Silence a compiler warning
The warning is a false positive on this line:

    if ((code = gs_glyphshow_begin(igs, glyph, imemory_local, &penum))
       < 0)

But glyph is initialised immediately above that:

    if ((code = glyph_show_setup(i_ctx_p, &glyph)) != 0)

So there is no way to reach that line without code being non-zero or
glyph being initialised. The compiler is simply confused by the return
macros in glyph_show_setup() I think.

Anyway, to get rid of the warning, simply initalise the variable.