Silence a compiler warning
This is actually a false positive, the code is:

    if ((code = glyph_show_setup(i_ctx_p, &glyph)) != 0)
        return code;
    if ((code = gs_glyphshow_begin(igs, glyph, imemory_local, &penum)) < 0)
        return code;

and the warning is against gs_glyphshow_begin. However glyph is actually
initalised in the line above by glyph_show_setup. There are paths to
exit that routine without setting glyph, but they all return non-zero
exit codes.

I suspect the return macros in glyph_show_setup and co are confusing
the compiler.

To silence the compiler warning, simply initialise the variable.