PDF interpreter - check the type of the Info dictionary before use
Bug #700820 " 	GPL Ghostscript 9.26: Unrecoverable error, exit code 1"

The problem here is that the trailer dictionary has an entry for the
/Info dictionary, but the object it poitns to is not a dictionary.

For most purposes this doesn't matter, but when the devce is pdfwrite
we try to maintain the information in that dictionary.

This commit tests the type of the object referenced by the key in the
trailer dictionary and if the object isn't a dictionary we don't try to
use it. We do flag an error though.