Bug 700815(1): Init gsubrNumberBias in Type 1 fonts and eexec decoder
The gsubrNumberBias is only used for Type 2 fonts, but the structure and
charstring interpreter are shared between Type 1 and 2 fonts, so we still need
to set gsubrNumberBias for Type 1 fonts (always 0 for Type 1).

Secondly, full initialise the eexec decoder state, along with some stack
variables during eexec decode setup.

This fixes most of the valgrind errors in Bug 700815, but not all.