Fix FillOrder tag for the TiffG3 and TiffG4 devices
Bug #700255 "Option FillOrder doesn't work anymore"

Commit 0ef309305417776a9f8b8cd143fbf3c5e89456b2 added the FillOrder
switch to the 'tiff' class of devices, but the fax Tiff devices already
processed that switch, and the commit did not remove the processing
from those devices.

This led to us trying to us ending up with the parameter in the params
list twice when we executed get_params (once from the tiff device
processing and once from the tiff fax device processing).

When we then next execute a 'put_param', the list has a duplicate
entry for FillOrder. We stop processing when we find the first entry,
which means the second entry never gets processed, leading to an
unused entry in the list, and that throws an error.

The simplest way to fix this is to remove the duplicate 'get_param()'
processing from the tiff fax devices, leaving it up to the generic
Tiff device.