Fix valgrind error with PXL patterns and high level devices
Bug #700573 " Uninitialised colors used in path plotting"

When the current colour space is a pattern, and the saved high level
colour space is also a pattern, we were checking every possible
component of ccolor.paint.values in both spaces against each other.

The problem is that in gx_hld_save_color we set almost all the values to
0, but then copy a single uninitialised value into the 0th index.

This leads to a valgrind error. Henry points out that the PostScript
interpreter has a hack specifically to work around this in zcolor.c,
zsetcolor() at around line 313.

This is wasteful, its far better (IMO) to rearrange the logic in
gx_hld_saved_color_equal(), we already bail out early if the colour
spaces are not equal. If the spaces are patterns, then there's no point
in comparing the device components, instead compare the pattern IDs
*before* we compare the device components, if they are the same then
return true, if they are different return false.

This avoids us having to check 2 sets of 64 values, and also avoids us
having to set 64 values every time we create a pattern.