Fixed/cleanup pdfi_create_indexed and add some warnings
There was some strange logic that always assumed there would be
PDF_INDIRECT in the colorspace dictionary, meaning some paths in the
code wouldn't be taken.  I'm not sure if there was a bug in practice,
but now that PDF_INDIRECT is handled by pdfi_array_get(), the logic was
defintely not going to work.  Rewrote this (mainly just changing the
if/then flow) and also cleaned up the error handling.  There were some
bugs in not properly cleaning up with the existing flow.

Also added a warning message for the case where the index array isn't
the size the code is expecting.  I expect there is something wrong with
the logic here.  I didn't change the logic, just added a warning for
the case that is currently aborting.

In pdf_image, changed the logic for when getting the colorspace fails,
to attempt to flush the image.  At least now it will print a warning saying
what is going on.

For the case of inline images with errors, I am not sure if it's
better to just abort the whole file. Right now it attempts to flush, which
may or may not work, and can result in reading garbage data.  Or maybe
there is a more robust way to do the flushing that doesn't rely on knowing
how big the image is?

I was working with the test file tests_private/comparefiles/H00216q.pdf,
which still doesn't work right, but is better than it was before.