Fix two reference counting problems.
When dealing with indirect object tokens 'R', we counted down the
'keyword' object, but then on every error path we counted it down again
which was wrong.

When dealing with 'inheritable' resource definitions when reading a
page dictionary we created the inheritable dictionary but didn't count
it up, so its reference count was always 0.