Fix indexed colour space handling
we were using sfgetc repeatedly to find the decompressed length of a
compressed lookup table. Unfortunately this doesn't work the way I would
exepct. For a 384 byte decompressed stream it stops returning bytes
after 360 bytes and returns '-1'.

I have a feeling we've run across this before, which is why we don't use
that function anywhere else.

Change it to use sfread instead. Works now.