pdfwrite - change the example PDF/A and PDF/X scripts
From Bug #700930 "Typo in example command for PDF/X-3 in VectorDevices.htm"

The reporter claimed that the (mis-spelled) command didn't work, which
was not surprising, since it contained a typo. After fixing the
documentation the reporter claimed that the command didn't work. This
is true and was missed because of the misleading title.

The PDFX_def.ps file was mistakenly pulling the ProcessColorModel from
systemdict, which would only be set if this was set on the command line.
The real problem is the fact that our setpagedevice/currentpagedevice
code doesn't truly reflect the device setup, so changes to the
ProcessColorModel caused by setting ColorConversionStrategy are not
reflected in the values returned by currentpagedevice.

This commit updates PDFX_def.ps to chaeck the ColorConversionStrategy
instead, which is more reliable, and then use the ProcessColorModel
from the currentpagedevice as a fallback if ColorConversionStrategy
is not set. (actually this should not be possible)

Update PDFA_def.ps similarly; although it was using the currentpagedvice
this is not a reliable method.

Finally; reformat the files, add some extra comments describing what the
code is doing, make it clearer that these files *should* be customised
and add additional some basic instruction on customising them.