Cleanup PDF_INDIRECT stuff
Make a pdfi_deref_loop_detect() that handles the loop detection along with
dereferencing, and change some code to call that.

Take out some direct handling of PDF_INDIRECT for some dictionary stuff,
since pdfi_dict_get() already handles it.

I believe this commit is a progression that fixes a few bugs, such as
on tests_private/pdf/sumatra/2339_-_dash_length_issues.pdf which used to
get an error on our version detecting a loop that I don't think is real.

This commit causes segfault on tests_private/pdf/sumatra/amino130M.pdf
on the cluster, but I looked at it and it seems memory is messed up
with or without this commit, and the commit just exposes the problem
differently.  I can get a segfault without this commit on args
"-Z: -dPDFDEBUG -sDEVICE=pkmraw -dMaxBitmap=0 -r300".
(So the segfault is probably a good thing that will make us look into this!)